requirements.txt
absl-py==0.12.0
astunparse==1.6.3
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cors==0.9.11
CRF==0.0.1a1
cymem==2.0.5
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==21.1.2
glob2==0.7
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
marisa-trie==0.7.5
Markdown==3.3.4
MarkupSafe==1.1.1
murmurhash==1.0.5
mysql-connector-python==8.0.23
nltk==3.6.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pathy==0.5.2
preshed==3.0.5
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
repoze.lru==0.7
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn-crfsuite==0.3.6
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==8.0.3
threadpoolctl==2.1.0
tldextract==3.1.0
tqdm==4.60.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
